GAN/requirements.txt
torch
torchvision
numpy
tqdm
hydra-core
wandb
